# api/openapi.yaml
openapi: 3.1.0
info:
  title: Collections Agent API
  version: 1.0.0
  description: REST API for the AI Receivables Collections Agent console, the bank webhook and admin controls. Serves the
    stories named in each operation's x-story-ids. Money fields are integer paise named *_paise (INR only); every amount shown
    to a user is formatted from these at the edge.
  contact:
    name: HACK-001 team
...
servers:
- url: http://localhost:8000/api/v1
  description: local Docker Compose
security:
- demoRole: []
...
paths:
...
  /customers/{id}/portal-link:
    post:
      operationId: createPortalLink
      summary: Create a 7-day portal link for one customer; token shown once
      tags:
      - portal
      x-story-ids:
      - HACK-003
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /customers/{id}/portal-links/revoke:
    post:
      operationId: revokePortalLinks
      summary: Revoke every live portal link of a customer
      tags:
      - portal
      x-story-ids:
      - HACK-003
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /portal/{token}:
    get:
      operationId: portalView
      summary: The customer's open invoices and promises; allow-listed fields only
      tags:
      - portal
      x-story-ids:
      - HACK-003
      x-roles: public (token)
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: token
        in: path
        required: true
        schema:
          type: string
          minLength: 20
          maxLength: 100
          pattern: ^[A-Za-z0-9_-]+$
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /portal/{token}/promise:
    post:
      operationId: portalPromise
      summary: Promise to pay up to the open balance within 60 days
      tags:
      - portal
      x-story-ids:
      - HACK-003
      x-roles: public (token)
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: token
        in: path
        required: true
        schema:
          type: string
          minLength: 20
          maxLength: 100
          pattern: ^[A-Za-z0-9_-]+$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - amount_paise
              - promised_date
              properties:
                amount_paise:
                  type: integer
                  minimum: 1
                promised_date:
                  type: string
                  format: date
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /portal/{token}/dispute:
    post:
      operationId: portalDispute
      summary: Raise a dispute on one of this customer's invoices
      tags:
      - portal
      x-story-ids:
      - HACK-003
      x-roles: public (token)
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: token
        in: path
        required: true
        schema:
          type: string
          minLength: 20
          maxLength: 100
          pattern: ^[A-Za-z0-9_-]+$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - invoice_number
              - description
              properties:
                invoice_number:
                  type: string
                  pattern: ^INV-\d+$
                description:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /portal/{token}/help:
    post:
      operationId: portalHelp
      summary: Ask for help; a collector is notified
      tags:
      - portal
      x-story-ids:
      - HACK-003
      x-roles: public (token)
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: token
        in: path
        required: true
        schema:
          type: string
          minLength: 20
          maxLength: 100
          pattern: ^[A-Za-z0-9_-]+$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - message
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /portal/{token}/pay:
    post:
      operationId: portalPay
      summary: SIMULATED Pay Now for one of this customer's invoices
      tags:
      - portal
      x-story-ids:
      - HACK-003
      x-roles: public (token)
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: token
        in: path
        required: true
        schema:
          type: string
          minLength: 20
          maxLength: 100
          pattern: ^[A-Za-z0-9_-]+$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - invoice_number
              properties:
                invoice_number:
                  type: string
                  pattern: ^INV-\d+$
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    demoRole:
      type: apiKey
      in: header
      name: X-Demo-Role
      description: admin, collector or viewer
    bankSignature:
      type: apiKey
      in: header
      name: X-Bank-Signature
      description: hex HMAC-SHA256 of '<X-Bank-Timestamp>.<raw body>' with BANK_WEBHOOK_SECRET
  parameters:
    ApiVersion:
      name: X-API-Version
      in: header
      required: false
      schema:
        type: string
        example: '2026-09-30'
...
  headers:
    XRequestId:
      schema:
        type: string
      description: Matches error.request_id and the server log line
  responses:
...
    Unauthorized:
      description: Not signed in
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Not signed in
              request_id: req_01J9Z6
    Forbidden:
      description: Signed in but not allowed
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: Signed in but not allowed
              request_id: req_01J9Z6
    NotFound:
      description: Not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Not found
              request_id: req_01J9Z6
...
    UnprocessableEntity:
      description: Valid shape, invalid content
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AMOUNT_MISMATCH
              message: Valid shape, invalid content
              request_id: req_01J9Z6
...
    InternalError:
      description: Unexpected error
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL
              message: Unexpected error
              request_id: req_01J9Z6
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable UPPER_SNAKE code, e.g. AMOUNT_MISMATCH, NOT_APPROVED, STALE_DRAFT
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  reason:
                    type: string
                required:
                - field
                - reason
            request_id:
              type: string
          required:
          - code
          - message
          - request_id
      required:
      - error
      example:
        error:
          code: AMOUNT_MISMATCH
          message: ₹4,50,000 is not the ledger amount for INV-1021
          details:
          - field: body
            reason: INV-1021 is 40000000 paise
          request_id: req_01J9Z6
...